Rename platform files to getAzuriteAccounts
