+++ sdk/tables/data-tables/src/TableServiceClient.ts
@@ -42,7 +42,7 @@ import { setTokenChallengeAuthenticationPolicy } from "./utils/challengeAuthenti
 import { tablesNamedKeyCredentialPolicy } from "#platform/tablesNamedCredentialPolicy";
 import { tablesSASTokenPolicy } from "./tablesSASTokenPolicy.js";
 import { tracingClient } from "./utils/tracing.js";
-import { isCosmosEndpoint } from "./utils/isCosmosEndpointUtil.js";
+import { isCosmosEndpoint } from "./utils/isCosmosEndpoint.js";
 
 /**
  * A TableServiceClient represents a Client to the Azure Tables service allowing you
