+++ sdk/tables/data-tables/src/utils/getAzuriteAccounts-browser.mts
