+++ sdk/tables/data-tables/src/utils/getAzuriteAccounts.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export function getAzuriteAccounts(): string | undefined {
+  return process.env.AZURITE_ACCOUNTS;
+}
